requirements.txt
alembic==1.4.3
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
coverage==5.3
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
python-decouple==3.3
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
